+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Fixes
 
 * Mask out all reserved bits in `set_bit_timing` before writing the register.
+### New Features
+
+* Implement the `embedded-can` 0.4.1 CAN traits.
 
 ## [0.7.0 - 2022-05-30](https://github.com/stm32-rs/bxcan/releases/tag/v0.7.0)